refactor: rename organizations dir
