+++ src/organization/index.ts
@@ -1,3 +1,3 @@
 export { Organization } from './entities/organization.entity';
-export { OrganizationsModule } from './organization.module';
+export { OrganizationModule } from './organization.module';
 export { OrganizationService } from './organization.service';
